Findbugs changes

+++ api/src/main/java/edu/cornell/mannlib/vitro/webapp/controller/freemarker/IndividualListController.java
@@ -250,7 +250,7 @@ public class IndividualListController extends FreemarkerHttpServlet {
         }else if ( requiredPages > MAX_PAGES && selectedPage > requiredPages - MAX_PAGES ){
             //the selected page is in the end of the list 
             int startPage = requiredPages - MAX_PAGES;      
-            double max = Math.ceil(size/pageSize);
+            double max = Math.ceil((double)size/(double)pageSize);
             for(int page = startPage; page <= max; page++ ){
                 records.add( new PageRecord( "page=" + page, Integer.toString(page), Integer.toString(page), selectedPage == page ) );            
             }
